+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 - BC BREAK - `Collection::toArray()` renamed to `Collection::all()`.
 - Added the factory methods `Collection::from()` and `Collection::empty()`.
 - Added helper methods `Collection::isEmpty()` and `Collection::isNotEmpty()`.
+- Added helper method `Collection::mapTo()` to map to a new Collection of a different type.
 - Added `Collection::reverse()` helper to reverse the order of the collection.
 - Support creating collections from Doctrine Collections using `Collection::fromDoctrine()`.
 - Support creating collections by mapping array of other objects using `Collection::mapFrom()`.